wmpower.py: Add a workaround for 2 different JSON syntax errors
Different versions of Whatminer firmware return corrupt JSON data.
This fix works only with a patch to the python whatsminer library that
returns the original string data in the exception. Of course a correct way
to fix this, would be to add a method to obtain the raw string data in the
first place, but this isn't currently possible.

Signed-off-by: David Jander <djander@gmail.com>