Add first test code of Rpi pico W based control module
This control module consists of a Raspberry pi pico W module with an small
OLED display, some buttons, 4 replays to control pumps and some NTC inputs
to monitor temperatures for a small immersion tank with second loop water
circulation. For now, this is only test code, to learn micropython, the
RP2040 and test the hardware.

Signed-off-by: David Jander <djander@gmail.com>