wmpower: Add rudimentary MQTT-based Home Assistant support
WM connection is managed and MQTT messages are published for HA auto-
configuration.
For now it presents a switch that permits to send "power_on" and
"power_off" commands to the miner, as well as 3 sensors for PSU power,
temperature and hash-rate. The SENSOR topic already contains a lot more
data, but HA config messages need to be added for them still.

Signed-off-by: David Jander <djander@gmail.com>