Controller2: Add one-off enable switch for power control
If the system starts and main power has not been enabled yet, wait for
user-interaction before attempting to enable power.
The UI will have a button that enables power control. This button is not
visible when power is already enabled.
The intention of this feature is to avoid any unsupervised power startup
on system boot if this program is started automatically as a systemd
service.

Signed-off-by: David Jander <djander@gmail.com>