Controller2: base_io.py: Bidir: Add nudge_left/nudge_right methods
Positioning of the bi-directional valve in center position is quite
critical. 1% error has a big impact on flow ratio.
Add these two methods to tweak the center position. For now it will be
used as a test. Later on, a control loop might be added to steer the
position more precisely.

Signed-off-by: David Jander <djander@gmail.com>